requirements.txt
Django==3.0
djangorestframework==3.12.4
flake8==4.0.1
isort==5.10.1
requests==2.28.0
python-dotenv==0.19.0
gunicorn==20.1.0
python-telegram-bot==13.7